Move image conversion code to the camera module
- Removes some hex-specific code
- Why? To attach the conversion code to the camera geometry
