--- ctapipe/image/geometry_converter.py
@@ -1,19 +1,0 @@
-"""
-collects converter functions from the various ``geometry_converter_XXX`` files in one
-common module
-"""
-from .geometry_converter_hex import (
-    convert_geometry_hex1d_to_rect2d,
-    convert_geometry_rect2d_back_to_hex1d,
-)
-from .geometry_converter_rect import (
-    convert_rect_image_1d_to_2d,
-    convert_rect_image_back_to_1d,
-)
-
-__all__ = [
-    "convert_geometry_hex1d_to_rect2d",
-    "convert_geometry_rect2d_back_to_hex1d",
-    "convert_rect_image_1d_to_2d",
-    "convert_rect_image_back_to_1d",
-]
